third_party/WebKit/LayoutTests/animations/usecounters/step-middle-timing-function-deprecated.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
'use strict';
// From web_feature.mojom
var kDeprecatedTimingFunctionStepMiddle = 2024;

test(() => {
  assert_false(internals.isUseCounted(document, kDeprecatedTimingFunctionStepMiddle));
  var element = document.createElement('div');
  element.style.animationTimingFunction = 'steps(3, middle)';
  document.documentElement.appendChild(element);
  assert_equals(getComputedStyle(element).animationTimingFunction, 'ease');
  assert_false(internals.isUseCounted(document, kDeprecatedTimingFunctionStepMiddle));
}, 'The step timing function with step position \'middle\' does not function outside of Web Animations.');

test(() => {
  assert_false(internals.isUseCounted(document, kDeprecatedTimingFunctionStepMiddle));
  document.documentElement.animate([], { easing: 'steps(3, middle)' });
  assert_true(internals.isUseCounted(document, kDeprecatedTimingFunctionStepMiddle));
}, 'The step timing function with step position \'middle\' is deprecated in Web Animations.');

</script>
